Add PDF-Tools
This is an intermediate modules for the reading and writing of PDF data.

It builds on PDF::Grammar adding serialization functions such as compression filters, encryption and indirect object fetching via index and writing of PDF files. It is capable of reading, editing and creation or incremental update of PDF files. Hopefully this will provide a useful set of classes and utilities for implementing higher level PDF modules.