Check for valid source URLs
This is one way to address the issue[^1] of wrong source urls in META
files. However, if there's github connectivity issue, we might get
false negatives and temporarily toss modules from ecosystem.

This needs to be done in modules.perl6.org codebase, so that the module
itself remains listed.

Stashing this commit for reuse there, but not for inclusion in this repo.

[1] https://github.com/perl6/modules.perl6.org/issues/74