Remove YAML from active exosystem
This module at first sight appears to provide YAML support to Raku
but on second sight, just provides an (incomplete and possibly
incorrect) way to dump a Raku object as YAML.  It hasn't had any
significant maintenance since 2010.

The repository has been archived.  The versions released so far
remain available in the REA.