Remove Raku::Parser from live ecosystem
With RakuAST becoming more and more mature, any work on interfacing
with old (Perl 6) based grammar and actions, is to become obsolete
pretty soon.  It really hurts to freeze this work by the late Jeff
Goff.  The repository will stay available as archived for any inspiration
seekers.  The released versions will remain available in the REA.