First stab at providing superseds mapping
This contains all of my modules that lived on PAUSE at some point.

Each line is supposed to consist of two or more "words".  The first
word is the original module:auth<…> string, and the next words are
the module:auth<…> strings that superseded the previous module:auth<…>.

The idea is to provide a community maintained source of superseding
information, so that sites such as raku.land can de-dupe distributions,
and other services can give a suggestion should a module no longer be
available/maintained.